requirements.txt
argparse~=1.4.0
Pillow~=9.0.1
torch~=1.11.0
numpy~=1.22.3
torchvision~=0.12.0
scikit-learn~=1.0.2
scipy~=1.8.0
matplotlib~=3.5.1